requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
bottle==0.13.2
cffi==1.17.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
loguru==0.7.3
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
proxy_tools==0.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyserial==3.5
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
setuptools==75.8.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==15.0.1
win32_setctime==1.2.0
